Refactor: “raw key 기반 structured UI”를 “semantic canonical key + legacy alias 호환 구조”로 끌어올리는 리팩터링
Changes to be committed:
	new file:   JSEngine/Asset/UI/PlaybackControls/PlayControlsJumpToNextKey.meta
	new file:   JSEngine/Asset/UI/PlaybackControls/PlayControlsJumpToPreviousKey.meta
	new file:   JSEngine/Asset/UI/PlaybackControls/PlayControlsLooping.meta
	new file:   JSEngine/Asset/UI/PlaybackControls/PlayControlsLoopingSelectionRange.meta
	new file:   JSEngine/Asset/UI/PlaybackControls/PlayControlsNoLooping.meta
	new file:   JSEngine/Asset/UI/PlaybackControls/PlayControlsPause.meta
	new file:   JSEngine/Asset/UI/PlaybackControls/PlayControlsPlayForward.meta
	new file:   JSEngine/Asset/UI/PlaybackControls/PlayControlsPlayReverse.meta
	new file:   JSEngine/Asset/UI/PlaybackControls/PlayControlsRecord.meta
	new file:   JSEngine/Asset/UI/PlaybackControls/PlayControlsSetPlaybackEnd.meta
	new file:   JSEngine/Asset/UI/PlaybackControls/PlayControlsSetPlaybackStart.meta
	new file:   JSEngine/Asset/UI/PlaybackControls/PlayControlsStop.meta
	new file:   JSEngine/Asset/UI/PlaybackControls/PlayControlsToEnd.meta
	new file:   JSEngine/Asset/UI/PlaybackControls/PlayControlsToFront.meta
	new file:   JSEngine/Asset/UI/PlaybackControls/PlayControlsToNext.meta
	new file:   JSEngine/Asset/UI/PlaybackControls/PlayControlsToPrevious.meta
	modified:   JSEngine/JSEngine.vcxproj
	modified:   JSEngine/JSEngine.vcxproj.filters
	modified:   JSEngine/Source/Editor/Animation/AnimationSequenceEditorDocument.cpp
	modified:   JSEngine/Source/Editor/Animation/AnimationSequenceEditorWidget.cpp
	modified:   JSEngine/Source/Editor/Animation/AnimationSequenceEditorWidget.h
	modified:   JSEngine/Source/Editor/Animation/AnimationSequenceNotifyPayloadSchema.cpp
	modified:   JSEngine/Source/Editor/Animation/AnimationSequenceNotifyPayloadSchema.h
	modified:   JSEngine/Source/Editor/Animation/AnimationSequenceNotifyValidation.cpp
	modified:   JSEngine/Source/Editor/UI/AnimInstanceStateSequenceContext.cpp
	modified:   JSEngine/Source/Engine/Animation/AnimInstance.cpp
	modified:   JSEngine/Source/Engine/Animation/AnimNotifyBuiltins.cpp
	modified:   JSEngine/Source/Engine/Animation/AnimNotifyPayloadParser.cpp
	modified:   JSEngine/Source/Engine/Animation/AnimNotifyPayloadParser.h
	new file:   JSEngine/Source/Engine/Animation/AnimNotifySemanticFieldNames.cpp
	new file:   JSEngine/Source/Engine/Animation/AnimNotifySemanticFieldNames.h

+++ JSEngine/Source/Engine/Animation/AnimInstance.cpp
@@ -514,7 +514,7 @@ void UAnimInstance::EvaluatePoseAtTime(const UAnimSequence* Sequence, float Curr
         }
     }
 
-    const float FrameRate = Model->FrameRate.AsDecimal();
+    const float FrameRate = static_cast<float>(Model->FrameRate.AsDecimal());
     if (FrameRate <= 0.0f)
     {
         return;
@@ -590,7 +590,7 @@ FQuat UAnimInstance::InterpolateKeys(const TArray<FQuat>& Keys, float Time, floa
 
 void UAnimInstance::BlendPoses(const FSkeletonPose& PoseA, const FSkeletonPose& PoseB, float BlendFactor, FSkeletonPose& OutPose)
 {
-    const int32 BoneCount = PoseA.LocalTransforms.size();
+    const int32 BoneCount = static_cast<int32>(PoseA.LocalTransforms.size());
     OutPose.LocalTransforms.resize(BoneCount);
 
     for (int32 i = 0; i < BoneCount; ++i)
@@ -615,7 +615,7 @@ float UAnimInstance::GetSequenceLength(const UAnimSequence* Sequence) const
         return 0.0f;
     }
 
-    const float FrameRate = Model->FrameRate.AsDecimal();
+    const float FrameRate = static_cast<float>(Model->FrameRate.AsDecimal());
     if (FrameRate <= 0.0f || Model->NumberOfKeys <= 1)
     {
         return 0.0f;
